refactor(synth-engine): consolidate duplicates and split params module
- Consolidate TWO_PI, wrap01, lerp into dsp_utils.rs (removed from 8 files)
- Delete unused oscillator.rs (duplicated dsp_utils content)
- Split params.rs (2411 lines) into params/ module directory:
  - envelopes.rs: EnvStep, StepEnvData
  - waveforms.rs: Algo, CzWaveform, BaseWaveform, WindowType
  - line.rs: LineParams, LineSelect, ModMode, PolyMode
  - modulation.rs: ModSource, ModDestination, ModMatrix
  - lfo.rs: LfoWaveform, LfoParams
  - filter.rs: FilterType, FilterParams
  - portamento.rs: PortamentoMode, PortamentoParams
  - fx_params.rs: All 17 FX param structs + FxSlotConfig
  - synth_params.rs: SynthParams, RandomParams, ModEnvParams
  - ui_meta.rs: UI metadata types, arrays, accessor functions
