+++ packages/cosmo-synth-engine/src/envelope.rs
@@ -3,6 +3,7 @@ use crate::envelope_map::{
     human_level_to_raw, human_rate_to_raw, raw_level_to_human, raw_rate_to_human,
 };
 use crate::params::{StepEnvData, SynthParams};
+use crate::dsp_utils::lerp;
 
 pub fn normalize_env_to_raw_if_human(kind: EnvelopeKind, env: &mut StepEnvData) {
     for step in env.steps.iter_mut() {
@@ -198,11 +199,6 @@ impl EnvGen {
 
         self.prev_level = self.output;
     }
-}
-
-#[inline]
-fn lerp(a: f32, b: f32, t: f32) -> f32 {
-    a + (b - a) * t
 }
 
 /// Converts a human rate [0..99] to a transition duration in seconds.
