+++ packages/cosmo-synth-engine/src/generators/karpunk.rs
@@ -3,6 +3,7 @@ extern crate alloc;
 use alloc::vec;
 use alloc::vec::Vec;
 
+use crate::dsp_utils::lerp;
 use crate::params::{Algo, EngineParamReadoutFormatV1};
 
 use super::{
@@ -272,11 +273,6 @@ pub fn blend(primary_algo: Algo, primary: f32, secondary: f32, blend: f32) -> f3
     } else {
         lerp(primary, secondary, blend)
     }
-}
-
-#[inline(always)]
-fn lerp(a: f32, b: f32, t: f32) -> f32 {
-    a + (b - a) * t
 }
 
 /// Simple LCG PRNG — produces a value in [-1.0, 1.0].
