+++ packages/cosmo-synth-engine/src/params/modulation.rs
@@ -1,0 +1,252 @@
+use serde::{Deserialize, Serialize};
+#[cfg(feature = "specta-bindings")]
+use specta::Type;
+
+/// Modulation source selector for modulation matrix routes.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, Default)]
+#[cfg_attr(feature = "specta-bindings", derive(Type))]
+#[serde(rename_all = "camelCase")]
+pub enum ModSource {
+    #[default]
+    Lfo1,
+    Lfo2,
+    Random,
+    ModEnv,
+    Velocity,
+    ModWheel,
+    Aftertouch,
+}
+
+/// Modulation destination selector for modulation matrix routes.
+#[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize, Default)]
+#[cfg_attr(feature = "specta-bindings", derive(Type))]
+#[serde(rename_all = "camelCase")]
+pub enum ModDestination {
+    #[default]
+    Volume,
+    Pitch,
+    Line1DcwBase,
+    Line1DcaBase,
+    Line1AlgoBlend,
+    Line2DetuneNote,
+    Line1Octave,
+    Line1AlgoParam1,
+    Line1AlgoParam2,
+    Line1AlgoParam3,
+    Line1AlgoParam4,
+    Line1AlgoParam5,
+    Line1AlgoParam6,
+    Line1AlgoParam7,
+    Line1AlgoParam8,
+    Line2DcwBase,
+    Line2DcaBase,
+    Line2AlgoBlend,
+    Line2DetuneFine,
+    Line2DetuneOctave,
+    Line2AlgoParam1,
+    Line2AlgoParam2,
+    Line2AlgoParam3,
+    Line2AlgoParam4,
+    Line2AlgoParam5,
+    Line2AlgoParam6,
+    Line2AlgoParam7,
+    Line2AlgoParam8,
+    FilterCutoff,
+    FilterResonance,
+    FilterEnvAmount,
+    ChorusMix,
+    DelayMix,
+    ReverbMix,
+    VibratoDepth,
+    VibratoRate,
+    IntPmRatio,
+    Line1DcoEnvStep1Level,
+    Line1DcoEnvStep1Rate,
+    Line1DcoEnvStep2Level,
+    Line1DcoEnvStep2Rate,
+    Line1DcoEnvStep3Level,
+    Line1DcoEnvStep3Rate,
+    Line1DcoEnvStep4Level,
+    Line1DcoEnvStep4Rate,
+    Line1DcoEnvStep5Level,
+    Line1DcoEnvStep5Rate,
+    Line1DcoEnvStep6Level,
+    Line1DcoEnvStep6Rate,
+    Line1DcoEnvStep7Level,
+    Line1DcoEnvStep7Rate,
+    Line1DcoEnvStep8Level,
+    Line1DcoEnvStep8Rate,
+    Line1DcwEnvStep1Level,
+    Line1DcwEnvStep1Rate,
+    Line1DcwEnvStep2Level,
+    Line1DcwEnvStep2Rate,
+    Line1DcwEnvStep3Level,
+    Line1DcwEnvStep3Rate,
+    Line1DcwEnvStep4Level,
+    Line1DcwEnvStep4Rate,
+    Line1DcwEnvStep5Level,
+    Line1DcwEnvStep5Rate,
+    Line1DcwEnvStep6Level,
+    Line1DcwEnvStep6Rate,
+    Line1DcwEnvStep7Level,
+    Line1DcwEnvStep7Rate,
+    Line1DcwEnvStep8Level,
+    Line1DcwEnvStep8Rate,
+    Line1DcaEnvStep1Level,
+    Line1DcaEnvStep1Rate,
+    Line1DcaEnvStep2Level,
+    Line1DcaEnvStep2Rate,
+    Line1DcaEnvStep3Level,
+    Line1DcaEnvStep3Rate,
+    Line1DcaEnvStep4Level,
+    Line1DcaEnvStep4Rate,
+    Line1DcaEnvStep5Level,
+    Line1DcaEnvStep5Rate,
+    Line1DcaEnvStep6Level,
+    Line1DcaEnvStep6Rate,
+    Line1DcaEnvStep7Level,
+    Line1DcaEnvStep7Rate,
+    Line1DcaEnvStep8Level,
+    Line1DcaEnvStep8Rate,
+    Line2DcoEnvStep1Level,
+    Line2DcoEnvStep1Rate,
+    Line2DcoEnvStep2Level,
+    Line2DcoEnvStep2Rate,
+    Line2DcoEnvStep3Level,
+    Line2DcoEnvStep3Rate,
+    Line2DcoEnvStep4Level,
+    Line2DcoEnvStep4Rate,
+    Line2DcoEnvStep5Level,
+    Line2DcoEnvStep5Rate,
+    Line2DcoEnvStep6Level,
+    Line2DcoEnvStep6Rate,
+    Line2DcoEnvStep7Level,
+    Line2DcoEnvStep7Rate,
+    Line2DcoEnvStep8Level,
+    Line2DcoEnvStep8Rate,
+    Line2DcwEnvStep1Level,
+    Line2DcwEnvStep1Rate,
+    Line2DcwEnvStep2Level,
+    Line2DcwEnvStep2Rate,
+    Line2DcwEnvStep3Level,
+    Line2DcwEnvStep3Rate,
+    Line2DcwEnvStep4Level,
+    Line2DcwEnvStep4Rate,
+    Line2DcwEnvStep5Level,
+    Line2DcwEnvStep5Rate,
+    Line2DcwEnvStep6Level,
+    Line2DcwEnvStep6Rate,
+    Line2DcwEnvStep7Level,
+    Line2DcwEnvStep7Rate,
+    Line2DcwEnvStep8Level,
+    Line2DcwEnvStep8Rate,
+    Line2DcaEnvStep1Level,
+    Line2DcaEnvStep1Rate,
+    Line2DcaEnvStep2Level,
+    Line2DcaEnvStep2Rate,
+    Line2DcaEnvStep3Level,
+    Line2DcaEnvStep3Rate,
+    Line2DcaEnvStep4Level,
+    Line2DcaEnvStep4Rate,
+    Line2DcaEnvStep5Level,
+    Line2DcaEnvStep5Rate,
+    Line2DcaEnvStep6Level,
+    Line2DcaEnvStep6Rate,
+    Line2DcaEnvStep7Level,
+    Line2DcaEnvStep7Rate,
+    Line2DcaEnvStep8Level,
+    Line2DcaEnvStep8Rate,
+    ChorusRate,
+    ChorusDepth,
+    DelayTime,
+    DelayFeedback,
+    DelayWarmth,
+    ReverbSpace,
+    ReverbPredelay,
+    ReverbDistance,
+    ReverbCharacter,
+    PhaserRate,
+    PhaserDepth,
+    PhaserFeedback,
+    PhaserMix,
+    Lfo1Rate,
+    Lfo1Depth,
+    Lfo1Symmetry,
+    Lfo1Offset,
+    Lfo2Rate,
+    Lfo2Depth,
+    Lfo2Symmetry,
+    Lfo2Offset,
+    RandomRate,
+    VibratoDelay,
+    CompressorThreshold,
+    CompressorRatio,
+    CompressorMakeup,
+    CompressorMix,
+    GrainDelayTime,
+    GrainDelayFeedback,
+    GrainDelayScatter,
+    GrainDelayDensity,
+    GrainDelayMix,
+    BitcrusherBits,
+    BitcrusherRateReduction,
+    BitcrusherMix,
+    ShimmerVerbShimmer,
+    ShimmerVerbSpace,
+    ShimmerVerbMix,
+    DistortionDrive,
+    DistortionTone,
+    DistortionMix,
+    JunoChorusMix,
+    RingModCarrierHz,
+    RingModMix,
+    TremoloRate,
+    TremoloDepth,
+    TremoloMix,
+    WavefolderDrive,
+    WavefolderFolds,
+    WavefolderMix,
+    LoFiDegrade,
+    LoFiWowDepth,
+    LoFiWowRate,
+    LoFiFlutterDepth,
+    LoFiFlutterRate,
+    LoFiTone,
+    LoFiMix,
+    EqGain80,
+    EqGain240,
+    EqGain750,
+    EqGain2200,
+    EqGain8000,
+}
+
+/// A single modulation route assignment.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[cfg_attr(feature = "specta-bindings", derive(Type))]
+#[serde(rename_all = "camelCase")]
+pub struct ModRoute {
+    pub source: ModSource,
+    pub destination: ModDestination,
+    pub amount: f32,
+    pub enabled: bool,
+}
+
+impl Default for ModRoute {
+    fn default() -> Self {
+        Self {
+            source: ModSource::Lfo1,
+            destination: ModDestination::Volume,
+            amount: 0.0,
+            enabled: false,
+        }
+    }
+}
+
+/// Collection of modulation routes.
+#[derive(Debug, Clone, Serialize, Deserialize, Default)]
+#[cfg_attr(feature = "specta-bindings", derive(Type))]
+#[serde(rename_all = "camelCase")]
+pub struct ModMatrix {
+    #[serde(default)]
+    pub routes: Vec<ModRoute>,
+}
